+++ core/insqypsr.c
@@ -565,8 +565,8 @@ static void ReplaceInstanceVariables(
            {
             bexp->type = FCALL;
             bexp->value = rindx_func;
-            eptr = GenConstant(theEnv,INTEGER_TYPE,CreateInteger(theEnv,(long long) ndepth));
-            eptr->nextArg = GenConstant(theEnv,INTEGER_TYPE,CreateInteger(theEnv,(long long) posn));
+            eptr = GenConstant(theEnv,INTEGER_TYPE,CreateInteger(theEnv,ndepth));
+            eptr->nextArg = GenConstant(theEnv,INTEGER_TYPE,CreateInteger(theEnv,posn));
             bexp->argList = eptr;
            }
          else if (sdirect == true)
@@ -640,9 +640,9 @@ static void ReplaceSlotReference(
             CloseStringSource(theEnv,"query-var");
             theExp->type = FCALL;
             theExp->value = func;
-            theExp->argList = GenConstant(theEnv,INTEGER_TYPE,CreateInteger(theEnv,(long long) ndepth));
+            theExp->argList = GenConstant(theEnv,INTEGER_TYPE,CreateInteger(theEnv,ndepth));
             theExp->argList->nextArg =
-              GenConstant(theEnv,INTEGER_TYPE,CreateInteger(theEnv,(long long) posn));
+              GenConstant(theEnv,INTEGER_TYPE,CreateInteger(theEnv,posn));
             theExp->argList->nextArg->nextArg = GenConstant(theEnv,TokenTypeToType(itkn.tknType),itkn.value);
             break;
            }
