+++ core/tmpltutl.h
@@ -65,11 +65,11 @@
    void                           MultiIntoSingleFieldSlotError(Environment *,struct templateSlot *,Deftemplate *);
    void                           CheckTemplateFact(Environment *,Fact *);
    bool                           CheckRHSSlotTypes(Environment *,struct expr *,struct templateSlot *,const char *);
-   struct templateSlot           *GetNthSlot(Deftemplate *,int);
+   struct templateSlot           *GetNthSlot(Deftemplate *,long long);
    int                            FindSlotPosition(Deftemplate *,CLIPSLexeme *);
    void                           PrintTemplateFact(Environment *,const char *,Fact *,bool,bool,const char *);
    void                           UpdateDeftemplateScope(Environment *);
-   struct templateSlot           *FindSlot(Deftemplate *,CLIPSLexeme *,short *);
+   struct templateSlot           *FindSlot(Deftemplate *,CLIPSLexeme *,unsigned short *);
    Deftemplate                   *CreateImpliedDeftemplate(Environment *,CLIPSLexeme *,bool);
 
 #endif /* _H_tmpltutl */
